Add files via upload
Somehow, the default info.json in the for-science QMK repositories resulted in an inverted slave (depending on which master side was plugged into the PC. This modified info.json file that I uploaded fixed that error. You can have a look into my keymap to see the current setting in this keyboard.